# backend/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "CreateMemberRequest": {
        "properties": {
          "email": {
            "format": "email",
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "maxLength": 200,
            "minLength": 12,
            "type": "string"
          },
          "role": {
            "enum": [
              "ADMIN",
              "MEMBER"
            ],
            "type": "string"
          }
        },
        "required": [
          "email",
          "role"
        ],
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "MemberView": {
        "properties": {
          "email": {
            "type": "string"
          },
          "membershipId": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "MembershipView": {
        "properties": {
          "membershipId": {
            "format": "uuid",
            "type": "string"
          },
          "organizationId": {
            "format": "uuid",
            "type": "string"
          },
          "organizationName": {
            "type": "string"
          },
          "organizationSlug": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "OrganizationView": {
        "properties": {
          "displayName": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "publicContact": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SessionView": {
        "properties": {
          "email": {
            "type": "string"
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/MembershipView"
            },
            "type": "array"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "UpdateMemberRequest": {
        "properties": {
          "role": {
            "enum": [
              "ADMIN",
              "MEMBER"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "ACTIVE",
              "DISABLED"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MiezMerker HTTP API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Email/password login, establishes a server-side session",
        "tags": [
          "auth-controller"
        ]
      }
    },
...
    "/api/v1/organizations/{organizationId}": {
      "get": {
        "operationId": "getOrganization",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Organization details; requires ACTIVE membership (tenant isolation)",
        "tags": [
          "organization-controller"
        ]
      }
    },
    "/api/v1/organizations/{organizationId}/members": {
      "get": {
        "operationId": "listMembers",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List members of my organization (ADMIN and MEMBER see the roster; cross-organization access is forbidden)",
        "tags": [
          "member-controller"
        ]
      },
      "post": {
        "operationId": "createMember",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "ADMIN creates a user and/or ACTIVE membership (test-phase bootstrap, no mail)",
        "tags": [
          "member-controller"
        ]
      }
    },
    "/api/v1/organizations/{organizationId}/members/{membershipId}": {
      "patch": {
        "operationId": "updateMember",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "membershipId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "ADMIN changes role or activates/disables a membership",
        "tags": [
          "member-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}